break out parse url
