+++ lib/index.js
@@ -8,9 +8,8 @@ import setupHtmlToArticleJson from 'html-to-article-json';
 import assert from 'assert';
 import setupArticle from 'article-json-html-render';
 import embeds from './embeds';
-import {parseInput as parseEmbeds} from 'embeds';
+import parseUrl from './parse-url';
 import setupEvents from './setup-events';
-import objectAssign from 'object-assign';
 
 const Article = setupArticle({
   embeds: embeds,
@@ -25,27 +24,6 @@ function parseEmbed (type) {
 
     return elm.__props__;
   };
-}
-
-function parseUrl (elm) {
-  if (!elm) {
-    return null;
-  }
-
-  const content = elm.textContent;
-  if (content.substring(0, 4) !== 'http') {
-    return null;
-  }
-
-  const parsed = parseEmbeds(content);
-  if (!parsed) {
-    return null;
-  }
-
-  return objectAssign({}, parsed, {
-    embedType: parsed.type,
-    type: 'embed'
-  });
 }
 
 const htmlToArticleJson = setupHtmlToArticleJson({
